requirements.txt
line-bot-sdk==2.0.1
boto3==1.20.9
Pillow==8.4.0
